Make ros1 version nodelet compatible.
